refactor(stella-cli): split the steering seams out of subsession.rs
Binding hooks onto the worker lane took `subsession.rs` from 1497 to 1506
lines, over the 1500-line ratchet and not grandfathered, so the file-size
guard failed. The baseline takes no new entry for a first-time crossing
(AGENTS.md "God files") — the remedy is a split.

`SteeringTap`, `MidTurnRoute` and `route_mid_turn` move to
`subsession/steering.rs`: one subject, where a prompt typed while a turn is
running goes and the tap that carries it. Every item in the moved block
already spelled its paths in full, so the new module needs no imports.
`subsession.rs` re-exports all three, leaving `subsession::route_mid_turn`
and `crate::subsession::SteeringTap` — the paths `command_deck.rs`,
`command_deck/settle.rs` and the subagent tests use — unchanged.

subsession.rs is 1373 lines after the move.
